requirements.txt
matplotlib==3.6.0
numpy==1.23.3
opencv-contrib-python==4.6.0.66
Pillow==9.1.0
scipy==1.9.1
